Update module structure as planned
