+++ lib/modules/team_stats.rb
@@ -1,45 +1,13 @@
 require_relative 'calculateable'
 require_relative 'gatherable'
 require_relative 'season_stats'
+require_relative 'summary'
 
 module TeamStats
   include Gatherable
   include Calculateable
   include SeasonStats
-
-  def favorite_opponent(team_id)
-    team_wins = get_wins_by_opponent(team_id)
-    team_games = get_total_games_by_opponent(team_id)
-    team_average_by_opponent = get_average_by_opponent(team_wins, team_games)
-
-    team_ident = team_average_by_opponent.max_by { |_opp_team, percent| percent }[0]
-    team_name = get_team_name_by_id(team_ident)
-
-    team_name
-  end
-
-  def rival(team_id)
-    team_wins = get_wins_by_opponent(team_id)
-    team_games = get_total_games_by_opponent(team_id)
-    team_average_by_opponent = get_average_by_opponent(team_wins, team_games)
-
-    team_ident = team_average_by_opponent.min_by { |_opp_team, percent| percent }[0]
-    team_name = get_team_name_by_id(team_ident)
-
-    team_name
-  end
-
-  def head_to_head(team_id)
-    team_wins = get_wins_by_opponent(team_id)
-    team_games = get_total_games_by_opponent(team_id)
-    team_average_by_opponent = get_average_by_opponent(team_wins, team_games)
-
-    avg_by_team_name = team_average_by_opponent.map do |key, value|
-      [get_team_name_by_id(key), value]
-    end.to_h
-
-    avg_by_team_name.sort.to_h
-  end
+  include Summary
 
   def get_total_games_by_opponent(team_id)
     @games.collection.inject(Hash.new(0)) do |team_games, game|
@@ -134,224 +102,6 @@ module TeamStats
 
   def goal_difference(hash, game)
     hash[game[0]] = (game[1].home_goals.to_i - game[1].away_goals.to_i).abs
-  end
-
-  def most_goals_scored(team_id)
-    team_goals = []
-    @games.collection.each do |game|
-      if game.last.home_team_id == team_id
-        team_goals << game.last.home_goals.to_i
-      elsif game.last.away_team_id == team_id
-        team_goals << game.last.away_goals.to_i
-      end
-    end
-    team_goals.max
-  end
-
-  def fewest_goals_scored(team_id)
-    team_goals = []
-    @games.collection.each do |game|
-      if game.last.home_team_id == team_id
-        team_goals << game.last.home_goals.to_i
-      elsif game.last.away_team_id == team_id
-        team_goals << game.last.away_goals.to_i
-      end
-    end
-    team_goals.min
-  end
-
-  def worst_loss(team_id)
-    team_goal_diff = []
-    @games.collection.each do |game|
-      if game.last.home_team_id == team_id && game.last.home_goals < game.last.away_goals
-        team_goal_diff << game.last.away_goals.to_i - game.last.home_goals.to_i
-      elsif game.last.away_team_id == team_id && game.last.away_goals < game.last.home_goals
-        team_goal_diff << game.last.home_goals.to_i - game.last.away_goals.to_i
-      end
-    end
-    team_goal_diff.max
-  end
-
-  def seasonal_summary(team_id)
-    reg_season = season_summary_regular_season_record(team_id)
-
-    regular_summary = reg_season.inject(Hash.new(Hash.new(0.0))) do |hash, record|
-      hash[record[0]] = {
-        regular_season: {
-          win_percentage: record[1][:win_percentage],
-          total_goals_scored: record[1][:goals],
-          total_goals_against: record[1][:goals_against],
-          average_goals_scored: record[1][:goals_scored_percentage],
-          average_goals_against: record[1][:goals_against_percentage]
-        }
-      }
-      hash
-    end
-
-    post_season = season_summary_postseason_record(team_id)
-    post_summary = post_season.inject(Hash.new(Hash.new(0.0))) do |hash, record|
-      hash[record[0]] = {
-        postseason: {
-          win_percentage: record[1][:win_percentage],
-          total_goals_scored: record[1][:goals],
-          total_goals_against: record[1][:goals_against],
-          average_goals_scored: record[1][:goals_scored_percentage],
-          average_goals_against: record[1][:goals_against_percentage]
-        }
-      }
-      hash
-    end
-
-    post_summary.merge(regular_summary) { |_season, post, reg| post.merge(reg) }
-  end
-
-  def season_summary_win_percentage(hash, season_id)
-    total_games = (hash[season_id][:win] + hash[season_id][:loss] + hash[season_id][:draw])
-    wins = hash[season_id][:win]
-    percentage = ((wins.to_f / total_games)).round(2)
-
-    if percentage.nan?
-      hash[season_id][:win_percentage] = 0.0
-    else
-      hash[season_id][:win_percentage] = percentage
-    end
-    hash
-  end
-
-  def season_summary_goals_percentage(hash, season_id)
-    total_games = (hash[season_id][:win] + hash[season_id][:loss] + hash[season_id][:draw])
-    goals = hash[season_id][:goals]
-    percentage = ((goals.to_f / total_games)).round(2)
-
-    if percentage.nan?
-      hash[season_id][:goals_scored_percentage] = 0.0
-    else
-      (hash[season_id][:goals_scored_percentage] = percentage)
-    end
-    hash
-  end
-
-  def season_summary_goals_against_percentage(hash, season_id)
-    total_games = (hash[season_id][:win] + hash[season_id][:loss] + hash[season_id][:draw])
-    goals = hash[season_id][:goals_against]
-    percentage = ((goals.to_f / total_games)).round(2)
-
-    if percentage.nan?
-      hash[season_id][:goals_against_percentage] = 0.0
-    else
-      (hash[season_id][:goals_against_percentage] = percentage)
-    end
-    hash
-  end
-
-  def season_summary_regular_season_record(team_id)
-    @seasons.teams[team_id].reduce({}) do |hash, season|
-      next(hash) if season[1].nil?
-      season_id = season[0]
-      team_season = season[1].flatten
-      hash[season[0]] = team_win_lose_draw_regular_season(team_id, team_season)
-      season_summary_win_percentage(hash, season_id)
-      season_summary_goals_percentage(hash, season_id)
-      season_summary_goals_against_percentage(hash, season_id)
-      hash
-    end
-  end
-
-  def season_summary_postseason_record(team_id)
-    @seasons.teams[team_id].reduce({}) do |hash, season|
-      next(hash) if season[1].nil?
-      season_id = season[0]
-      team_season = season[1].flatten
-      hash[season[0]] = team_win_lose_draw_postseason(team_id, team_season)
-      season_summary_win_percentage(hash, season_id)
-      season_summary_goals_percentage(hash, season_id)
-      season_summary_goals_against_percentage(hash, season_id)
-      hash
-    end
-  end
-
-  def team_win_lose_draw_regular_season(team_id, team_season)
-    record = { win: 0, loss: 0, draw: 0, regular_season_games: 0, win_percentage: 0, goals: 0, goals_against: 0 }
-    team_season.reduce({}) do |hash, game|
-      if team_id == game.home_team_id && (game.home_goals > game.away_goals) && game.type == 'Regular Season'
-        record[:win] += 1
-        record[:regular_season_games] += 1
-        record[:goals] += game.home_goals.to_i
-        record[:goals_against] += game.away_goals.to_i
-      elsif team_id == game.home_team_id && (game.home_goals < game.away_goals) && game.type == 'Regular Season'
-        record[:loss] += 1
-        record[:regular_season_games] += 1
-        record[:goals] += game.home_goals.to_i
-        record[:goals_against] += game.away_goals.to_i
-      elsif team_id == game.away_team_id && (game.away_goals > game.home_goals) && game.type == 'Regular Season'
-        record[:win] += 1
-        record[:regular_season_games] += 1
-        record[:goals] += game.away_goals.to_i
-        record[:goals_against] += game.home_goals.to_i
-      elsif team_id == game.away_team_id && (game.away_goals < game.home_goals) && game.type == 'Regular Season'
-        record[:loss] += 1
-        record[:regular_season_games] += 1
-        record[:goals] += game.away_goals.to_i
-        record[:goals_against] += game.home_goals.to_i
-      elsif game.home_goals == game.away_goals && game.type == 'Regular Season'
-        record[:draw] += 1
-        record[:regular_season_games] += 1
-        record[:goals] += game.away_goals.to_i if team_id == game.away_team_id
-        record[:goals] += game.home_goals.to_i if team_id == game.home_team_id
-        record[:goals_against] += game.away_goals.to_i if team_id == game.home_team_id
-        record[:goals_against] += game.home_goals.to_i if team_id == game.away_team_id
-      end
-      hash = record
-      hash
-    end
-  end
-
-  def team_win_lose_draw_postseason(team_id, team_season)
-    record = { win: 0, loss: 0, draw: 0, postseason_games: 0, win_percentage: 0.0, goals: 0, goals_against: 0 }
-    team_season.reduce({}) do |hash, game|
-      if team_id == game.home_team_id && (game.home_goals > game.away_goals) && game.type == 'Postseason'
-        record[:win] += 1
-        record[:postseason_games] += 1
-        record[:goals] += game.home_goals.to_i
-        record[:goals_against] += game.away_goals.to_i
-      elsif team_id == game.home_team_id && (game.home_goals < game.away_goals) && game.type == 'Postseason'
-        record[:loss] += 1
-        record[:postseason_games] += 1
-        record[:goals] += game.home_goals.to_i
-        record[:goals_against] += game.away_goals.to_i
-      elsif team_id == game.away_team_id && (game.away_goals > game.home_goals) && game.type == 'Postseason'
-        record[:win] += 1
-        record[:postseason_games] += 1
-        record[:goals] += game.away_goals.to_i
-        record[:goals_against] += game.home_goals.to_i
-      elsif team_id == game.away_team_id && (game.away_goals < game.home_goals) && game.type == 'Postseason'
-        record[:loss] += 1
-        record[:postseason_games] += 1
-        record[:goals] += game.away_goals.to_i
-        record[:goals_against] += game.home_goals.to_i
-      elsif game.home_goals == game.away_goals && game.type == 'Postseason'
-        record[:draw] += 1
-        record[:postseason_games] += 1
-        record[:goals] += game.away_goals.to_i if team_id == game.away_team_id
-        record[:goals] += game.home_goals.to_i if team_id == game.home_team_id
-        record[:goals_against] += game.away_goals.to_i if team_id == game.home_team_id
-        record[:goals_against] += game.home_goals.to_i if team_id == game.away_team_id
-      end
-      hash = record
-      hash
-    end
-  end
-
-  def best_season(team_id)
-    final_percentage = team_season_wins_hash(team_id)
-    answer = final_percentage.max_by{|season, win_percentage| win_percentage}
-    answer[0]
-  end
-
-  def worst_season(team_id)
-    final_percentage = team_season_wins_hash(team_id)
-    answer = final_percentage.min_by{|season, win_percentage| win_percentage}
-    answer[0]
   end
 
   def team_season_wins_hash(team_id)
