+++ lib/stat_tracker.rb
@@ -9,4 +9,86 @@ class StatTracker < Tracker
   include GameStats
   include SeasonStats
   include TeamStats
+
+  def favorite_opponent(team_id)
+    team_wins = get_wins_by_opponent(team_id)
+    team_games = get_total_games_by_opponent(team_id)
+    team_average_by_opponent = get_average_by_opponent(team_wins, team_games)
+
+    team_ident = team_average_by_opponent.max_by { |_opp_team, percent| percent }[0]
+    team_name = get_team_name_by_id(team_ident)
+
+    team_name
+  end
+
+  def rival(team_id)
+    team_wins = get_wins_by_opponent(team_id)
+    team_games = get_total_games_by_opponent(team_id)
+    team_average_by_opponent = get_average_by_opponent(team_wins, team_games)
+
+    team_ident = team_average_by_opponent.min_by { |_opp_team, percent| percent }[0]
+    team_name = get_team_name_by_id(team_ident)
+
+    team_name
+  end
+
+  def head_to_head(team_id)
+    team_wins = get_wins_by_opponent(team_id)
+    team_games = get_total_games_by_opponent(team_id)
+    team_average_by_opponent = get_average_by_opponent(team_wins, team_games)
+
+    avg_by_team_name = team_average_by_opponent.map do |key, value|
+      [get_team_name_by_id(key), value]
+    end.to_h
+
+    avg_by_team_name.sort.to_h
+  end
+
+  def most_goals_scored(team_id)
+    team_goals = []
+    @games.collection.each do |game|
+      if game.last.home_team_id == team_id
+        team_goals << game.last.home_goals.to_i
+      elsif game.last.away_team_id == team_id
+        team_goals << game.last.away_goals.to_i
+      end
+    end
+    team_goals.max
+  end
+
+  def fewest_goals_scored(team_id)
+    team_goals = []
+    @games.collection.each do |game|
+      if game.last.home_team_id == team_id
+        team_goals << game.last.home_goals.to_i
+      elsif game.last.away_team_id == team_id
+        team_goals << game.last.away_goals.to_i
+      end
+    end
+    team_goals.min
+  end
+
+  def worst_loss(team_id)
+    team_goal_diff = []
+    @games.collection.each do |game|
+      if game.last.home_team_id == team_id && game.last.home_goals < game.last.away_goals
+        team_goal_diff << game.last.away_goals.to_i - game.last.home_goals.to_i
+      elsif game.last.away_team_id == team_id && game.last.away_goals < game.last.home_goals
+        team_goal_diff << game.last.home_goals.to_i - game.last.away_goals.to_i
+      end
+    end
+    team_goal_diff.max
+  end
+
+  def best_season(team_id)
+    final_percentage = team_season_wins_hash(team_id)
+    answer = final_percentage.max_by{|season, win_percentage| win_percentage}
+    answer[0]
+  end
+
+  def worst_season(team_id)
+    final_percentage = team_season_wins_hash(team_id)
+    answer = final_percentage.min_by{|season, win_percentage| win_percentage}
+    answer[0]
+  end
 end
